requirements.txt
pygame==1.9.3
configparser==3.5.0
http://antlr3.org/download/Python/antlr_python_runtime-3.1.3.tar.gz
https://netix.dl.sourceforge.net/project/pyfuzzy/pyfuzzy/pyfuzzy-0.1.0/pyfuzzy-0.1.0.tar.gz

PyYAML==5.4.1
typing==3.10.0.0
Shapely==1.7.1
